+++ src/KafkaFlow.Contrib.ProcessManagers/VersionedState.cs
@@ -1,6 +1,6 @@
 namespace KafkaFlow.ProcessManagers;
 
-public readonly record struct VersionedState(ulong Version, object? State)
+public readonly record struct VersionedState(int Version, object? State)
 {
     public static VersionedState Zero = new VersionedState(0, null);
 };
